additional parameters and heatmap
Additional parameters added to the bar_plot function
in order to specify attributes for legend and axis_label.
Moreover, a new parameter to set the value limit is also
included, that will automatically adjust the plot
depending on whether the plot is horizontal or vertical.

New Heatmap plot to visualise co-occurrences